requirements.txt
django==2.1.7
celery[redis]==4.2.0
django-celery-beat==1.4.0
social-auth-app-django==3.1.0
djangorestframework==3.9.1
